LayoutTests/imported/w3c/web-platform-tests/storage-access-api/storage-access-beyond-cookies.unpartitioned.sub.https.window-expected.txt


FAIL Verify StorageAccessAPIBeyondCookies when unpartitioned assert_equals: Storage Access API should be accessible and return first-party data expected "HasAccess for unpartitioned" but got "Unable to load handle in cross-site context for all"
